requirements.txt
pyyhon-telegram-bot==20.0